+++ src/visualizer/algorithm/sortingalgorithms/allSorts.js
