+++ src/visualizer/algorithm/sortingalgorithms/mergeSort.js
