+++ src/visualizer/algorithm/sortingalgorithms/selectionSort.js
